requirements.txt
aniso8601==7.0.0
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
email_validator==2.1.2
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-gql==0.1.0
fastgraphql==0.0.12
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==3.2.3
graphql-relay==2.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.5
promise==2.3
pycparser==2.22
pydantic==1.10.16
pydantic_core==2.18.4
Pygments==2.18.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
Rx==1.6.3
shellingham==1.5.4
singledispatch==3.7.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
starlette==0.37.2
strawberry-graphql==0.235.0
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.20.0
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
